refactor: isolate vegetation builders
