+++ lib/Structures.lua
@@ -50,6 +50,7 @@ local Buildings = V.require("Buildings")
 local TileShape = V.require("TileShape")
 local Budget = V.require("BuildBudget")
 local GridKey = V.require("GridKey")
+local VegetationBuilder = V.require("VegetationBuilder")
 
 local Structures = {}
 
@@ -503,10 +504,10 @@ function Structures.forMap(map)
     -- ---- tall grass: two standing tuft rows per tile. BODY only: the 2D
     -- renderer never draws a neighbour's ring, and standing scenery past a
     -- map's edge would poke into the map next door ----
-    Structures.buildGrass(S, map, 0, tw - 1, 0, th - 1, data)
+  VegetationBuilder.buildGrass(S, map, 0, tw - 1, 0, th - 1, data)
 
     -- ---- flowers: the animated meadow tile stands as a 1px cutout ----
-    Structures.buildFlowers(S, map, tw, th, x0, x1, y0, y1, data)
+  VegetationBuilder.buildFlowers(S, map, tw, th, x0, x1, y0, y1, data)
   end
 
   -- ---- authored ground under pinned props ----
@@ -3461,414 +3462,15 @@ end
 -- out again in texture space -- so a run that is six pixels wide in the
 -- union may be two pixels wide in the frame on screen, and the four pixels
 -- that dropped out took the union's end walls with them. What is left
--- exposed is an interior boundary, which had no wall because in the union
--- it was not a boundary at all. That is the gap that survived closing the
--- run ends: the first frame looked solid and every other frame did not.
---
--- So an animated standee gets a wall on BOTH sides of EVERY pixel. A wall
--- between two lit pixels is enclosed by the front and back faces and never
--- seen; the moment its neighbour is keyed out it becomes the edge, already
--- in place and already wearing the right colour. Each is inset a hair into
--- its own pixel so the two that meet at a boundary are not coplanar -- the
--- voxel pass draws with culling off, and two quads in the same plane would
--- z-fight rather than politely take turns.
-local SIDE_INSET = 0.03
-
-local function sideQuads(quads, ix, ix2, yBot, yTop, zB, zF,
-                         ax0, ay0, atlasW, atlasH, py, lit, everyPixel)
-  local function texel(px)
-    return (ax0 + px + 0.5) / atlasW, (ay0 + py + 0.5) / atlasH
-  end
-  local function left(px, at)
-    local u, v = texel(px)
-    quads[#quads + 1] = {                 -- facing -X
-      { at, yBot, zB }, { at, yBot, zF },
-      { at, yTop, zF }, { at, yTop, zB },
-      uv = { { u, v }, { u, v }, { u, v }, { u, v } },
-      shade = OBJ_SHADE.side,
-    }
-  end
-  local function right(px, at)
-    local u, v = texel(px)
-    quads[#quads + 1] = {                 -- facing +X
-      { at, yBot, zF }, { at, yBot, zB },
-      { at, yTop, zB }, { at, yTop, zF },
-      uv = { { u, v }, { u, v }, { u, v }, { u, v } },
-      shade = OBJ_SHADE.side,
-    }
-  end
-  if everyPixel then
-    for px = ix, ix2 do
-      left(px, px + SIDE_INSET)
-      right(px, px + 1 - SIDE_INSET)
-    end
-    return
-  end
-  if not lit(ix - 1, py) then left(ix, ix) end
-  if not lit(ix2 + 1, py) then right(ix2, ix2 + 1) end
-end
-
--- A tall-grass CELL is four tufts: 2x2 tiles, and each 8x8 tile is one
--- whole clump of grass. Each tile stands as its own thin per-pixel slab
--- at ITS OWN depth -- the cell's north tile row in the north half of the
--- cell, the south row in the south half -- over the flat grass base the
--- tile already renders. So the player walks BETWEEN the two rows, and
--- the southern row occludes their feet the way the 2D grass overdraw
--- did. Transparency respected: only the tuft strokes stand. Runs of
--- adjacent pixels merge into single quads, and one template per grass
--- tile id is stamped across the map (grass comes in fields).
---
--- One tile is ONE standing piece, full height. The first cut split each
--- tile again into its top and bottom four art rows and stood those at
--- two different depths, which cut every blade that runs down the tile
--- clean in half -- the two halves ended up 4px tall and 4px apart in
--- depth, so a clump read as two stubs rather than one tuft.
-local GRASS_THICK = 2
-
-local function grassTemplate(map, data, tileId)
-  local perRow = map.tileset.tilesPerRow or 16
-  local atlasW = map.tileset.imageWidth or 128
-  local atlasH = map.tileset.imageHeight or 48
-  local ax0 = (tileId % perRow) * 8
-  local ay0 = math.floor(tileId / perRow) * 8
-
-  local function opaque(px, py)
-    if px < 0 or px > 7 or py < 0 or py > 7 then return false end
-    local r, g, b, a = data:getPixel(ax0 + px, ay0 + py)
-    return a > 0 and math.min(r, g, b) <= 0.83
-  end
-
-  local quads = {}
-  -- the slab stands across the middle of its own tile, so the two tile
-  -- rows of a cell are half a cell apart in depth
-  local zMid = 4
-  local zB, zF = zMid - GRASS_THICK / 2, zMid + GRASS_THICK / 2
-  for iy = 0, 7 do
-    local yTop = 8 - iy
-    local yBot = yTop - 1
-    local ix = 0
-    while ix < 8 do
-      if opaque(ix, iy) then
-        local ix2 = ix
-        while ix2 + 1 < 8 and opaque(ix2 + 1, iy) do
-          ix2 = ix2 + 1
-        end
-        local u0 = (ax0 + ix + 0.05) / atlasW
-        local u1 = (ax0 + ix2 + 0.95) / atlasW
-        local v0 = (ay0 + iy + 0.05) / atlasH
-        local v1 = (ay0 + iy + 0.95) / atlasH
-        quads[#quads + 1] = {           -- front
-          { ix, yBot, zF }, { ix2 + 1, yBot, zF },
-          { ix2 + 1, yTop, zF }, { ix, yTop, zF },
-          uv = { { u0, v1 }, { u1, v1 }, { u1, v0 }, { u0, v0 } },
-          shade = 1,
-        }
-        quads[#quads + 1] = {           -- back
-          { ix2 + 1, yBot, zB }, { ix, yBot, zB },
-          { ix, yTop, zB }, { ix2 + 1, yTop, zB },
-          uv = { { u1, v1 }, { u0, v1 }, { u0, v0 }, { u1, v0 } },
-          shade = 0.68,
-        }
-        -- blade tips: a top strip where the row above is clear
-        if not opaque(ix, iy - 1) then
-          quads[#quads + 1] = {
-            { ix, yTop, zB }, { ix2 + 1, yTop, zB },
-            { ix2 + 1, yTop, zF }, { ix, yTop, zF },
-            uv = { { u0, v0 }, { u1, v0 }, { u1, v0 }, { u0, v0 } },
-            shade = 1,
-          }
-        end
-        -- and underneath, where a blade ends in mid-air over the ground
-        if not opaque(ix, iy + 1) then
-          quads[#quads + 1] = {
-            { ix, yBot, zF }, { ix2 + 1, yBot, zF },
-            { ix2 + 1, yBot, zB }, { ix, yBot, zB },
-            uv = { { u0, v1 }, { u1, v1 }, { u1, v1 }, { u0, v1 } },
-            shade = OBJ_SHADE.bottom,
-          }
-        end
-        -- and the run's two end walls, which is what makes a blade a solid
-        -- thing rather than two billboards you can see between (sideQuads
-        -- above argues it, and why each wall wears its end pixel's colour)
-        sideQuads(quads, ix, ix2, yBot, yTop, zB, zF,
-                  ax0, ay0, atlasW, atlasH, iy, opaque)
-        ix = ix2 + 1
-      else
-        ix = ix + 1
-      end
-    end
-  end
-  return quads
-end
-
+-- Compatibility façade: vegetation ownership lives in VegetationBuilder.
 function Structures.buildGrass(S, map, x0, x1, y0, y1, data)
-  local templates = {}
-  local quads = S.grassQuads
-  for ty = y0, y1 do
-    for tx = x0, x1 do
-      Budget.tick()
-      local k = GridKey.of(tx, ty)
-      local s = S.shapeAt[k]
-      -- tufts only where the CELL is tall grass by the engine's own rule
-      -- (isGrassCell: the cell's collision tile). The grass GRAPHIC also
-      -- appears as decorative filler inside ordinary ground blocks, and a
-      -- tile-level test sprouted tufts all over town plazas.
-      if s and s.art == "grass"
-         and map:isGrassCell(math.floor(tx / 2), math.floor(ty / 2)) then
-        local tileId = S.tileAt[k]
-        local tpl = templates[tileId]
-        if not tpl then
-          tpl = grassTemplate(map, data, tileId)
-          templates[tileId] = tpl
-        end
-        local wx, wz = tx * 8, ty * 8
-        for _, q in ipairs(tpl) do
-          quads[#quads + 1] = {
-            { q[1][1] + wx, q[1][2], q[1][3] + wz },
-            { q[2][1] + wx, q[2][2], q[2][3] + wz },
-            { q[3][1] + wx, q[3][2], q[3][3] + wz },
-            { q[4][1] + wx, q[4][2], q[4][3] + wz },
-            uv = q.uv, shade = q.shade,
-          }
-        end
-      end
-    end
-  end
-end
-
--- ---- flowers ----
-
--- The animated flower tile stands up as a billboard ONE VOXEL deep, cut
--- to the drawing's darkest tones PLUS everything they enclose -- the
--- round-scenery hull's rule: flood the tile border through every
--- non-dark pixel, and what the flood cannot reach is the flower, its
--- pale petal insides included. The mesh is static and the flower is
--- not, so the geometry spans the UNION of that mask over the base art
--- and every animation frame, and TerrainAtlas rewrites the tile's slot
--- each step with only the CURRENT frame's mask opaque -- the rest keyed
--- to alpha, which the voxel shader discards. The standing silhouette
--- trims itself frame by frame in texture space; the sway animates
--- without a vertex moving, off the same engine clock as the flat path.
---
--- The ground beneath is synthesized from the commonest flat neighbour,
--- like the ground under a detected prop: the tile's own slot no longer
--- holds art anyone can draw flat.
-local FLOWER_THICK = 1
-
-local function flowerFrames(tileset, tileId)
-  local out = {}
-  local ok, declared = pcall(function()
-    if tileset.animatedTiles then return tileset.animatedTiles end
-    local TileRenderer = require("src.render.TileRenderer")
-    return TileRenderer.defaultAnimatedTiles(tileset)
-  end)
-  if not ok then return out end
-  for _, spec in ipairs(type(declared) == "table" and declared or {}) do
-    if spec.kind == "frames" and spec.tile == tileId then
-      for _, path in pairs(spec.images or {}) do
-        local okF, frame = pcall(Assets.imageData, path)
-        if okF and frame then out[#out + 1] = frame end
-      end
-    end
-  end
-  return out
-end
-
-local function flowerTemplate(map, data, tileId)
-  local tileset = map.tileset
-  local perRow = tileset.tilesPerRow or 16
-  local atlasW = tileset.imageWidth or 128
-  local atlasH = tileset.imageHeight or 48
-  local ax0 = (tileId % perRow) * 8
-  local ay0 = math.floor(tileId / perRow) * 8
-
-  -- per image: dark tones, then the border flood that finds what they
-  -- enclose. Each image closes over ITS OWN outline before the union --
-  -- a pocket two frames only enclose together is not part of either.
-  local dark = {}
-  local function markMask(img, ox, oy)
-    local d, reach, stack = {}, {}, {}
-    for py = 0, 7 do
-      for px = 0, 7 do
-        local r, g, b, a = img:getPixel(ox + px, oy + py)
-        if a > 0 and math.min(r, g, b) <= 0.5 then
-          d[py * 8 + px] = true
-        end
-      end
-    end
-    for i = 0, 7 do
-      for _, s in ipairs({ i, 56 + i, i * 8, i * 8 + 7 }) do
-        if not d[s] and not reach[s] then
-          reach[s] = true
-          stack[#stack + 1] = s
-        end
-      end
-    end
-    while #stack > 0 do
-      local p = table.remove(stack)
-      local px, py = p % 8, math.floor(p / 8)
-      for _, dir in ipairs(DIRS4) do
-        local nx, ny = px + dir[1], py + dir[2]
-        if nx >= 0 and nx < 8 and ny >= 0 and ny < 8 then
-          local ni = ny * 8 + nx
-          if not d[ni] and not reach[ni] then
-            reach[ni] = true
-            stack[#stack + 1] = ni
-          end
-        end
-      end
-    end
-    for i = 0, 63 do
-      if d[i] or not reach[i] then dark[i] = true end
-    end
-  end
-  markMask(data, ax0, ay0)
-  for _, frame in ipairs(flowerFrames(tileset, tileId)) do
-    pcall(markMask, frame, 0, 0)
-  end
-
-  local function on(px, py)
-    if px < 0 or px > 7 or py < 0 or py > 7 then return false end
-    return dark[py * 8 + px] == true
-  end
-
-  local quads = {}
-  local zB = 4 - FLOWER_THICK / 2      -- one slab at the tile's middle
-  local zF = zB + FLOWER_THICK
-  for py = 0, 7 do
-    Budget.tick()
-    local yTop, yBot = 8 - py, 7 - py
-    local ix = 0
-    while ix < 8 do
-      if on(ix, py) then
-        local ix2 = ix
-        while ix2 + 1 < 8 and on(ix2 + 1, py) do ix2 = ix2 + 1 end
-        local u0 = (ax0 + ix + 0.05) / atlasW
-        local u1 = (ax0 + ix2 + 0.95) / atlasW
-        local v0 = (ay0 + py + 0.05) / atlasH
-        local v1 = (ay0 + py + 0.95) / atlasH
-        quads[#quads + 1] = {           -- front
-          { ix, yBot, zF }, { ix2 + 1, yBot, zF },
-          { ix2 + 1, yTop, zF }, { ix, yTop, zF },
-          uv = { { u0, v1 }, { u1, v1 }, { u1, v0 }, { u0, v0 } },
-          shade = OBJ_SHADE.front,
-        }
-        quads[#quads + 1] = {           -- back
-          { ix2 + 1, yBot, zB }, { ix, yBot, zB },
-          { ix, yTop, zB }, { ix2 + 1, yTop, zB },
-          uv = { { u1, v1 }, { u0, v1 }, { u0, v0 }, { u1, v0 } },
-          shade = OBJ_SHADE.back,
-        }
-        -- ------- the shell, closed on all four remaining faces
-        --
-        -- A flower SWAYS: the geometry spans the union of every animation
-        -- frame's mask and each frame is cut back out of it in texture
-        -- space (see the header). So "is there a pixel next door" has two
-        -- different answers -- one in the union this mesh was built from,
-        -- and one in the frame actually on screen -- and only the second
-        -- decides what is exposed.
-        --
-        -- Closing the union's own edges is therefore not enough, and was
-        -- the bug the first cut of this shipped: the base frame looked
-        -- solid and every other frame still had gaps, because a pixel that
-        -- drops out of a frame takes the union's wall with it and leaves an
-        -- interior boundary that never had one.
-        --
-        -- So every pixel gets a cap on all four of its remaining faces,
-        -- whatever its neighbours do. A cap between two lit pixels sits
-        -- inside the slab, enclosed by the front and back faces, and is
-        -- never seen; the moment its neighbour is keyed out it IS the edge,
-        -- already there and already wearing the right colour. Each samples
-        -- its own pixel's texel, so it appears and vanishes with the pixel
-        -- it belongs to rather than with the one it is closing off.
-        --
-        -- Inset a hair into its own pixel, because the voxel pass draws
-        -- with culling off: the two caps that meet at a boundary would be
-        -- coplanar and z-fight rather than politely take turns.
-        for px = ix, ix2 do
-          local tu = (ax0 + px + 0.5) / atlasW
-          local tv = (ay0 + py + 0.5) / atlasH
-          local xa, xb = px, px + 1
-          local yT = yTop - SIDE_INSET
-          local yB = yBot + SIDE_INSET
-          quads[#quads + 1] = {           -- the pixel's own lid
-            { xa, yT, zB }, { xb, yT, zB }, { xb, yT, zF }, { xa, yT, zF },
-            uv = { { tu, tv }, { tu, tv }, { tu, tv }, { tu, tv } },
-            shade = OBJ_SHADE.top,
-          }
-          quads[#quads + 1] = {           -- and its floor
-            { xa, yB, zF }, { xb, yB, zF }, { xb, yB, zB }, { xa, yB, zB },
-            uv = { { tu, tv }, { tu, tv }, { tu, tv }, { tu, tv } },
-            shade = OBJ_SHADE.bottom,
-          }
-        end
-        sideQuads(quads, ix, ix2, yBot, yTop, zB, zF,
-                  ax0, ay0, atlasW, atlasH, py, on, true)
-        ix = ix2 + 1
-      else
-        ix = ix + 1
-      end
-    end
-  end
-  return quads
+  return VegetationBuilder.buildGrass(S, map, x0, x1, y0, y1, data)
 end
 
 function Structures.buildFlowers(S, map, tw, th, x0, x1, y0, y1, data)
-  local templates = {}
-  -- flowerQuads, not objectQuads: flowers sit on WALKABLE cells, so
-  -- their mesh draws after the characters with the character pull
-  -- (ChunkMesher's flower mesh) -- terrain-baked they lose the depth
-  -- fight against the pulled card whenever the player stands among them
-  local quads = S.flowerQuads
-  for ty = y0, y1 do
-    for tx = x0, x1 do
-      Budget.tick()
-      local k = GridKey.of(tx, ty)
-      local s = S.shapeAt[k]
-      if s and s.art == "flower" then
-        -- the tile's atlas slot carries only the standing cutout now, so
-        -- EVERY flower position -- ring included -- paints synthesized
-        -- ground instead of its own art: the commonest flat neighbour
-        -- that is not itself a flower, else the map's commonest ground
-        -- (forMap's end-of-build vote resolves the `false`)
-        S.skip[k] = true
-        local votes, best, bestN = {}, nil, 0
-        for _, d in ipairs(DIRS4) do
-          local nk = GridKey.of(tx + d[1], ty + d[2])
-          local ns = S.shapeAt[nk]
-          if ns and ns.flat and ns.class ~= "void"
-             and ns.class ~= "flower" then
-            local t = S.tileAt[nk]
-            votes[t] = (votes[t] or 0) + 1
-            if votes[t] > bestN then best, bestN = t, votes[t] end
-          end
-        end
-        S.ground[k] = best or false
-
-        -- standee BODY only, like grass: standing scenery past a map's
-        -- edge would poke into the map next door
-        if tx >= 0 and ty >= 0 and tx < tw and ty < th then
-          local tileId = S.tileAt[k]
-          local tpl = templates[tileId]
-          if not tpl then
-            tpl = flowerTemplate(map, data, tileId)
-            templates[tileId] = tpl
-          end
-          local wx, wz = tx * 8, ty * 8
-          for _, q in ipairs(tpl) do
-            quads[#quads + 1] = {
-              { q[1][1] + wx, q[1][2], q[1][3] + wz },
-              { q[2][1] + wx, q[2][2], q[2][3] + wz },
-              { q[3][1] + wx, q[3][2], q[3][3] + wz },
-              { q[4][1] + wx, q[4][2], q[4][3] + wz },
-              uv = q.uv, shade = q.shade,
-            }
-          end
-        end
-      end
-    end
-  end
+  return VegetationBuilder.buildFlowers(S, map, tw, th, x0, x1, y0, y1, data)
 end
+
 
 -- Drop one map's analysis (Cut changed the block layer) or everything.
 -- Hull templates key on art content (tileset + tiles), which a block edit
